requirements.txt
pillow==10.3.0
python-dotenv==1.0.1
